+++ python/paddle/__init__.py
@@ -44,6 +44,7 @@ import paddle.metric
 import paddle.device
 import paddle.regularizer
 import paddle.incubate
+import paddle.autograd
 
 # TODO: define alias in tensor and framework directory
